requirements.txt
anthropic>=0.40.0          # Anthropic SDK (compatible with ZhipuAI Coding Plan)
openai>=1.0.0              # OpenAI SDK (for MiniMax compatibility)
prompt-toolkit>=3.0.0      # Terminal interaction
python-dotenv>=1.0.0        # Environment variable management
rich>=13.0.0              # Terminal beautiful output
